Feat: Report relatedType 변경
int -> Long

+++ src/main/java/com/hamss2/KINO/api/entity/Report.java
@@ -21,9 +21,9 @@ public class Report {
     @Column(nullable = false)
     private String content;
 
-    private int reportType;
+    private Integer reportType;
 
-    private int relatedType;
+    private Long relatedType;
 
     private Long relatedId;
